Simplify control flow in make_retrostep()
Method make_retrostep() of Chemical class was returning a set of reaction
SMILES or None, if applying transform resulted in empty product list. I rewrite
it to make it return just an empty set in case of failure. It allowed to
simplify a bit of control flow and should have no unwanted side-effects..