Optimze checking if a reaction is in the database
Previously, to check if an incoming reaction is published (i.e. is in the
database) a separate query (is_published() function) had to be executed. To
make the matter worse, the query was using unindexed field ('smiles') making
those query very time consuming. I removed this in efficient function and
changed the logic behind the search.

Now, the known reactions are added to the pool of all reactions, and
appropriately marked, on the fly while processing a chemical by utilizing
indexed reaction ids stored in the document database describing a chemical
('reaction' field).