Allow user to specify RNG seed explicitly
Previous version of Sample class does not allow to specify explicitly the seed
of the pseudo-random number generator (RNG). Thus it was not possible to select
exactly the same sample from the database other than to read it from an
external file.

I added an option 'rng_seed' which can be pass during class instantiation
ensuring that for the same values of 'rng_seed' that the same sample will be
return, providing that the database has not changed.