Improve decriptors' calculations
The get_descriptors() method of the Reaction class had a long, hard to
maintain list of descriptor calls.

I added an iterator over methods of Chemical class related to
calculating descriptors. It allowed me to replace the long list of
'manual' calls of descriptor methods in Reaction class with a simple and
concise loop.

Additionally, I made the method to exclude a simple compounds having a
single-atom H-depleted molecular graph (e.g. water with its 'O') from
calculations as they adjacency and distance matrices are equal to zero
which confuse many topological descriptors.