Do not include not recreated published reactions
Previous version was including *all* published reaction from the
database to the final dataset, even those which was not or cannot be
recreated by the available transforms, e.g. multi-product reactions. It
violated the assumption that the published reactions are a subset of
possible reactions generated by the transforms.

I modified the code to make it exclude published reactions not recreated
by the transform from the reaction list for which the descriptors are
calculated.