Replace comparison operators by dedicated method
The way we consider reactions to be equal -- one reaction's
reactants/products are to be others subsets -- is NOT symmetric
relation. It depends on order of the operands. Thus, using '==' operator
for checking it could be counterintuitive and leads to errors as one
usually expected that the order of such a comparison does not depend on
arguments order. Thus I replaced the operators with dedicated method
'equals' which clearly indicates which set inclusions we are interested
in.