Remove SMILES difference tracking
In case of published reactions which were regenerated by retrosynthetic
transforms the script was storing both reaction SMILES, the original and
recreated one. As we do not need them at the moment, I removed the code
responsible for doing this.

For future reference , it is better to save this information in class
instances and not in the main script as it clutters in unnecessarily.