Add reaction selection
Calculating reaction descriptors is becoming a bottleneck for large
number of reactions. Also, we are mainly interested in comparing scores
of reactions sharing the same products. Thus, I implemented a simple
selection scheme which discards newly generated reactions if their
products are not available through published ones. Moreover, for each
products having both categories of reactions, only a pair of published
and unpublished reaction is selected.