Introduce popularity attribute
In general, more than one transform can be applied to the same product.
Some of them can have little practical meaning. After all, the algorithm
only checks whether transform's retrons can be match to target molecule.
In an attempt to differentiate transforms, and in turn reactions,
according to their likelihood I added a new attribute to Transform and
Reaction classes -- popularity.  It indicates how many times a given
transform was successfully applied to a sample from a database of
chemical reactions. Current transform popularity values are stored as a
dictionary in 'popularity.py'.